警告対応
(1) PacketCount 型を明示的に 64 ビットで宣言

+++ Include/TsSplitter/Common/TsSplitterTypes.h
@@ -120,7 +120,7 @@ typedef     BtByte  *           LpByteWriteBuf;
 **    パケット数。
 **/
 
-typedef     size_t              PacketCount;
+typedef     uint64_t            PacketCount;
 
 /**
 **    サービス ID
